src/backend/requirements.txt
alembic==1.13.2
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
typing_extensions==4.12.2
Werkzeug==3.0.4
pytest
python-dotenv==0.19.2